backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
cryptography==41.0.7
mysqlclient==2.2.0
stripe==7.7.0
django-environ==0.11.2
django-cors-headers==4.3.1
